v 1.6

+++ controller/comment.js
@@ -3,15 +3,16 @@ const CommentLikes = require("../classes/CommentLikes/CommentLikes")
 
 const CommentFunc = require('../classes/Comment/Function')
 const FriendshipFunc = require("../classes/Friendship/Function")
-const Constant = require("../classes/Comment/Constant")
+const Constant = require("../classes/Comment/Constant").Constant
 const utils = require("../utils")
-const TEXT = require("../text").TEXT
+const ERRORS = require("../errorConstants").ERRORS
 
 /** 
  * creates new comment 
  * @param {object} req 
  * @param {object} res 
  * @param {Object} next
+ * @returns 
  */
 exports.createComment = async(req, res, next) => {
     let reqTweetid = parseInt(req.body.tid)
@@ -26,7 +27,7 @@ exports.createComment = async(req, res, next) => {
         let tidCheckResponse = await Comment.checkTweetIdExists(reqTweetid)
         let tidCheck = tidCheckResponse.data
         if(tidCheck.length == 0){
-            return utils.sendResponse(res, false, {}, TEXT.noTweetExist)
+            return utils.sendResponse(res, false, {},ERRORS.noTweetExist)
     }
     } catch (error) {
         return utils.sendResponse(res, false, {}, error)
@@ -35,32 +36,32 @@ exports.createComment = async(req, res, next) => {
     try {
         let uidCheckResponse = await Comment.checkUserIdExists(reqUserid)
         if(uidCheckResponse.data.length == 0){
-            return utils.sendResponse(res, false, {}, TEXT.noUserExists)
+            return utils.sendResponse(res, false, {}, ERRORS.noUserExists)
     }
     } catch (error) {
         return utils.sendResponse(res, false, {}, error)
     }
     
     try {
-        let newCommentResponse = await Comment.createComment(req.body.tid, req.userid, req.body.comment)
+        let newCommentResponse = await Comment.createComment(reqTweetid, reqUserid, reqComment)
         return utils.sendResponse(res, true, newCommentResponse.data, "")
     } catch (error) {
         return utils.sendResponse(res, false, {}, error)
     }
 }
 
-
 /**
- * Reads all comment
+ * Reads (gets) all comment
  * @param {Object} req 
  * @param {Object} res 
  * @param {Object} next
+ * @returns 
  */
 exports.readComment = async(req, res, next) => {
     try {
-        let reqTweetid = parseInt(req.body.tid)
-        if(!reqTweetid){
-            return utils.sendResponse(res, false, {}, TEXT.noTweetId)
+        let reqTweetid = parseInt(req.headers.tid)
+        if(!reqTweetid || isNaN(reqTweetid)){
+            return utils.sendResponse(res, false, {}, ERRORS.noTweetId)
         }
         let commentsResponse = await Comment.readComment(reqTweetid)
         return utils.sendResponse(res, true, commentsResponse.data, "")
@@ -69,12 +70,12 @@ exports.readComment = async(req, res, next) => {
     }
 }
 
-
 /**
  * Update Comment 
  * @param {Object} req 
  * @param {Object} res 
  * @param {Object} next
+ * @returns 
  */
 exports.updateComment = async(req, res, next) => {
     try {
@@ -82,8 +83,8 @@ exports.updateComment = async(req, res, next) => {
         let reqTweetid = parseInt(req.body.tid)
         let reqUserid = parseInt(req.userid)
         let reqComment = (req.body.comment || "").toString()
-        if(!reqId || !reqTweetid || !reqUserid || !reqComment){
-            return utils.sendResponse(res, false, {}, TEXT.someFieldsMissing)
+        if(!reqId || isNaN(reqId) || !reqTweetid || isNaN(reqTweetid) || !reqUserid || isNaN(reqUserid) || !reqComment){
+            return utils.sendResponse(res, false, {}, ERRORS.someFieldsMissing)
         }
         let updateCommentResponse = await Comment.updateComment(reqId, reqTweetid, reqUserid, reqComment)
         return utils.sendResponse(res, true, updateCommentResponse.data, "")
@@ -97,14 +98,15 @@ exports.updateComment = async(req, res, next) => {
  * @param {Object} req 
  * @param {Object} res 
  * @param {Object} next
+ * @returns 
  */
 exports.deleteComment = async(req, res, next) => {
     try {
         let reqId = parseInt(req.body.id)
         let reqTweetid = parseInt(req.body.tid)
         let reqUserid = parseInt(req.userid)
-        if(!reqId || !reqTweetid || !reqUserid){
-            return utils.sendResponse(res, false, {}, TEXT.someFieldsMissing)
+        if(!reqId || isNaN(reqId) || !reqTweetid || isNaN(reqTweetid) || !reqUserid || isNaN(reqUserid)){
+            return utils.sendResponse(res, false, {}, ERRORS.someFieldsMissing)
         }
         let deleteCommentResponse = await Comment.deleteComment(reqId, reqTweetid, reqUserid)
         return utils.sendResponse(res, true, deleteCommentResponse.data, "")
@@ -113,12 +115,12 @@ exports.deleteComment = async(req, res, next) => {
     }
 }
 
-
 /**
  * Gets all comment under all tweets
  * @param {Object} req 
  * @param {Object} res 
  * @param {Object} next
+ * @returns 
  */
 //All in tweets
 exports.allInComments = async(req, res, next) => {
@@ -136,17 +138,18 @@ exports.allInComments = async(req, res, next) => {
  * @param {Object} req 
  * @param {Object} res 
  * @param {Object} next
+ * @returns 
  */
 exports.commentsUnderTweets = async(req, res, next) => {
     try {
-        let pageno = parseInt(req.params.pageno)
-        if(!pageno){
+        let pageno = parseInt(req.headers.pageno)
+        if(!pageno || isNaN(pageno)){
             pageno = Constant.defaultPageNo
         }
-        let reqTweetid = parseInt(req.params.tweetid)
+        let reqTweetid = parseInt(req.headers.tweetid)
         let reqUserid = parseInt(req.userid)
-        if(!reqTweetid || !reqUserid){
-            return utils.sendResponse(res, false, {}, TEXT.someFieldsMissing)
+        if(!reqTweetid || isNaN(reqTweetid) || !reqUserid || isNaN(reqUserid)){
+            return utils.sendResponse(res, false, {}, ERRORS.someFieldsMissing)
         }
         pageno = pageno-1;
         let pageSize = Constant.defaultPageSize;
