+++ controller/commentLikes.js
@@ -2,21 +2,20 @@ const CommentLikes = require("../classes/CommentLikes/CommentLikes")
 
 const utils = require("../utils")
 const ERRORS = require("../errorConstants").ERRORS
-const TEXT = require("../text").TEXT
-
 
 /**
  * Entry for one like, creates a comment like entry and likes a comment
  * @param {Object} req 
  * @param {Object} res 
  * @param {Object} next
+ * @returns 
  */
 exports.createLike = async (req, res, next) => {
     let reqUserId = parseInt(req.userid)
-    let reqcommentId = parseInt(req.params.commentid)
-    let reqLikeType = (req.params.likeType || "").toString()
-    if(!reqUserId || !reqcommentId || !reqLikeType){
-        return utils.sendResponse(res, false, {}, TEXT.someFieldsMissing)
+    let reqcommentId = parseInt(req.body.commentid)
+    let reqLikeType = (req.body.likeType || "").toString()
+    if(!reqUserId || isNaN(reqUserId) || !reqcommentId || isNaN(reqcommentId) || !reqLikeType){
+        return utils.sendResponse(res, false, {}, ERRORS.someFieldsMissing)
     }
     let isalreadyLiked = await CommentLikes.readLike(reqUserId, reqcommentId)
     if(isalreadyLiked.success == false){
@@ -31,12 +30,13 @@ exports.createLike = async (req, res, next) => {
  * @param {Object} req 
  * @param {Object} res 
  * @param {Object} next
+ * @returns 
  */
 exports.readLike = async (req, res, next) => {
     let reqUserId = parseInt(req.userid)
-    let reqcommentId = parseInt(req.params.commentid)
-    if(!reqUserId || !reqcommentId){
-        return utils.sendResponse(res, false, {}, TEXT.someFieldsMissing)
+    let reqcommentId = parseInt(req.headers.commentid)
+    if(!reqUserId || isNaN(reqUserId) || !reqcommentId || isNaN(reqcommentId)){
+        return utils.sendResponse(res, false, {}, ERRORS.someFieldsMissing)
     }
     let readcomment = await CommentLikes.readLike(reqUserId, reqcommentId)
     if(readcomment.success == false){
@@ -50,12 +50,13 @@ exports.readLike = async (req, res, next) => {
  * @param {Object} req 
  * @param {Object} res 
  * @param {Object} next
+ * @returns 
  */
 exports.deleteLike = async (req, res, next) => {
     let reqUserId = parseInt(req.userid)
-    let reqcommentId = parseInt(req.params.commentid)
-    if(!reqUserId || !reqcommentId){
-        return utils.sendResponse(res, false, {}, TEXT.someFieldsMissing)
+    let reqcommentId = parseInt(req.body.commentid)
+    if(!reqUserId || isNaN(reqUserId) || !reqcommentId || isNaN(reqcommentId)){
+        return utils.sendResponse(res, false, {}, ERRORS.someFieldsMissing)
     }
     let deleteLike = await CommentLikes.deleteLike(reqUserId, reqcommentId)
     return utils.sendResponse(res, true, deleteLike.data, "")
@@ -66,11 +67,12 @@ exports.deleteLike = async (req, res, next) => {
  * @param {Object} req 
  * @param {Object} res 
  * @param {Object} next
+ * @returns 
  */
 exports.whoAllLikedcomment = async (req, res, next) => {
-    let reqcommentId = parseInt(req.params.commentid)
-    if(!reqcommentId){
-        return utils.sendResponse(res, false, {}, TEXT.someFieldsMissing)
+    let reqcommentId = parseInt(req.headers.commentid)
+    if(!reqcommentId || isNaN(reqcommentId)){
+        return utils.sendResponse(res, false, {}, ERRORS.someFieldsMissing)
     }
     let whoAllLikedcomment = await CommentLikes.whoAllLikedcomment(reqcommentId)
     return utils.sendResponse(res, true, whoAllLikedcomment.data, "")
